services/python-scraper/requirements/base.txt
# Base requirements for FastAPI and core functionality
fastapi==0.115.14
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1

# Database drivers
pymongo==4.13.2
motor==3.7.1  # Async MongoDB driver
redis==6.2.0

# HTTP client and utilities
httpx==0.28.1
python-multipart==0.0.20

# Image processing
Pillow==10.4.0

# Authentication
PyJWT==2.10.1
python-jose[cryptography]==3.5.0

# Environment and configuration
python-dotenv==1.1.1

# Logging and monitoring
structlog==25.4.0
